update license; minor improvements
