+++ docs/metadata/unique.md
@@ -1,4 +1,4 @@
-# unique
+# Unique
 
 ## name
