+++ lib/dialects/mysql/index.js
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2019-present Andrey Vereshchak
+ * Copyright (c) 2019-present Andrew Vereshchak
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
@@ -8,7 +8,7 @@
 
 const TableInterface = require('./tables');
 const SequenceInterface = require('./sequences');
-const ConnectionManager = require('./connectionManager');
+const ConnectionManager = require('./connection-manager');
 
 /**
  *
