+++ lib/dialects/mysql/tables/index.js
@@ -1,26 +1,26 @@
 /**
- * Copyright (c) 2019-present Andrey Vereshchak
+ * Copyright (c) 2019-present Andrew Vereshchak
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  */
 'use strict';
 
-const queries = require('./queries');
+const QueryGenerator = require('./query-generator');
 const parser = require('./parser');
 const utils = require('../../../utils');
 
 class TableReader {
   static getColumns(metalize, names) {
     return metalize._client
-      .query(queries.getColumns(names))
+      .query(QueryGenerator.getColumns(names))
       .then(utils.group('table'))
       .then(utils.map(parser.columns));
   }
 
   static getConstraints(metalize, names) {
     return metalize._client
-      .query(queries.getConstraints(names))
+      .query(QueryGenerator.getConstraints(names))
       .then(utils.group('table'))
       .then(
         utils.map(utils.pipe(parser.constraints, utils.cleanableGroup('type')))
@@ -29,7 +29,7 @@ class TableReader {
 
   static getIndexes(metalize, names) {
     return metalize._client
-      .query(queries.getIndexes(names))
+      .query(QueryGenerator.getIndexes(names))
       .then(utils.group('table'))
       .then(utils.map(parser.indexes));
   }
