+++ lib/dialects/postgres/connection-manager.js
@@ -1,12 +1,11 @@
 /**
- * Copyright (c) 2019-present Andrey Vereshchak
+ * Copyright (c) 2019-present Andrew Vereshchak
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  */
 'use strict';
 
-const helpers = require('../../helpers');
 const BaseConnectionManager = require('../base/connectionManeger');
 
 let lib;
@@ -23,9 +22,11 @@ class ConnectionManager extends BaseConnectionManager {
     if (_options.client) {
       this._client = _options.client;
     } else {
-      lib = lib || this._loadDialectModule('mysql2/promise');
-      this._connect = async () => {
-        this._client = await lib.createConnection(_options.connectionConfig);
+      lib = lib || this._loadDialectModule('pg');
+      this._connect = () => {
+        const { Client } = lib;
+        this._client = new Client(_options.connectionConfig);
+        return this._client.connect();
       };
       this._end = () => {
         const client = this._client;
@@ -35,22 +36,11 @@ class ConnectionManager extends BaseConnectionManager {
     }
   }
 
-  async query(sql) {
+  query(sql) {
     if (!this._client) {
-      await this._connect();
+      this._connect();
     }
-    return this._client
-      .query(sql)
-      .then(([rows]) => rows)
-      .then(rows => {
-        if (rows) {
-          return Array.isArray(rows)
-            ? rows.map(helpers.lowercaseKeys)
-            : helpers.lowercaseKeys(rows);
-        } else {
-          return rows;
-        }
-      });
+    return this._client.query(sql);
   }
 }
