+++ lib/dialects/postgres/sequences/index.js
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2019-present Andrey Vereshchak
+ * Copyright (c) 2019-present Andrew Vereshchak
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
@@ -7,14 +7,14 @@
 'use strict';
 
 const parser = require('./parser');
-const queries = require('./queries');
+const QueryGenerator = require('./query-generator');
 const utils = require('../../../utils');
 const BaseSequenceReader = require('../../base/sequence');
 
 class SequenceReader extends BaseSequenceReader {
   static getProperties(metalize, names) {
     return metalize._client
-      .query(queries.getSequence(names))
+      .query(QueryGenerator.getSequence(names))
       .then(utils.get('rows'))
       .then(sequences =>
         sequences.reduce((acc, properties) => {
