+++ lib/dialects/postgres/tables/index.js
@@ -1,19 +1,19 @@
 /**
- * Copyright (c) 2019-present Andrey Vereshchak
+ * Copyright (c) 2019-present Andrew Vereshchak
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  */
 'use strict';
 
-const queries = require('./queries');
+const QueryGenerator = require('./query-generator');
 const parser = require('./parser');
 const utils = require('../../../utils');
 
 class TableReader {
   static getColumns(metalize, names) {
     return metalize._client
-      .query(queries.getColumns(names))
+      .query(QueryGenerator.getColumns(names))
       .then(utils.get('rows'))
       .then(utils.group('table'))
       .then(utils.map(parser.columns));
@@ -21,7 +21,7 @@ class TableReader {
 
   static getConstraints(metalize, names) {
     return metalize._client
-      .query(queries.getConstraints(names))
+      .query(QueryGenerator.getConstraints(names))
       .then(utils.get('[1].rows')) // `1` is the result of executing sql-query to get constraints
       .then(utils.group('table'))
       .then(
@@ -33,7 +33,7 @@ class TableReader {
 
   static getIndexes(metalize, names) {
     return metalize._client
-      .query(queries.getIndexes(names))
+      .query(QueryGenerator.getIndexes(names))
       .then(utils.get('rows'))
       .then(utils.group('table'))
       .then(utils.map(parser.indexDefinitions));
@@ -41,7 +41,7 @@ class TableReader {
 
   static async filterExistingTables(metalize, names) {
     const exists = await Promise.all(
-      names.map(name => metalize._client.query(queries.tableExist(name)))
+      names.map(name => metalize._client.query(QueryGenerator.tableExist(name)))
     );
     return names.filter((_, index) =>
       utils.get([index, 'rows', 0, 'exists'], exists)
