+++ lib/helpers.js
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2019-present Andrey Vereshchak
+ * Copyright (c) 2019-present Andrew Vereshchak
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
