+++ package.json
@@ -49,8 +49,7 @@
   },
   "husky": {
     "hooks": {
-      "pre-commit": "lint-staged",
-      "pre-push": "npm run test"
+      "pre-commit": "npm run test && lint-staged"
     }
   },
   "lint-staged": {
